+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Features:
 
 * extended PDDL support (`imply` expressions)
 * improved command-line interface
+* new command-line option `--log-level` to control which status messages should be shown
+* new command-line option `--warnings-as-errors` to abort program execution upon warnings (replaces `--warning-level`)
 
 ## 3.0.3 (2016-09-02)